vm: add skeleton userspace virtual memory server
The systask is named "vm" to avoid confusion between sys/sysmem and a
*sys/mem.

Launching the server involves a fair amount of protocol between root, sysmem,
and vm; and initialization needs more still, so this changeset is accordingly
huge. On the upside, starting "vm" as an ordinary systask allows use of the
full sys/crt capabilities, threading and all; so the elaborate rigamarole is
certainly worth it.

This changeset also reduces sysmem's initial allocation because it tends to
get most physical memory that'd've overlapped vm's own virtual memory. For
more, implement transfers from vm to sysmem as pressure increases, or memory
compression, or something.

The changeset also adds pinning of boot modules so that they don't get blown
away during sigma0 pump. That should be noted, since the root changes border
enormous to begin with.