sys/crt: store thread control data below stack bottom
This removes the per-process thread hash and the question of its locking,
instead using ExchangeRegisters to do tid-to-thread lookups.

While a syscall is more expensive than a local hashtable lookup, it's
balanced out by the opposite (and much more common) lookup having changed
to a pointer-transform job instead of a hash table query.