define and implement sysinfopage, add process exits
The system information page is a special bit of memory in the root task which
maps it to vm which maps it to a myriad userspace processes. The origin page
is therefore pinned to the root task in sysmem so that it doesn't get
replaced, which would invalidate vm's intermediate mapping and all the child
maps. Sysmem::alter_flags() was added to achieve this; it allows setting the
SMATTR_PIN flag which causes sysmem replacement to age the page but never
unmap it.

This changeset also adds an exit(3) analogue to Sneks::Proc and uses it to
cleanly exit the spawn partner program. That test reads all green afterward.