root: send maps from the sysmem pager, not grants
Otherwise it'd wind up in a condition where a page had been granted because of
swap activity, but send_phys_to_sysmem() would still be trying to grant it a
second time. Thing is, on the second go-around granting the same address means
granting zeroes, which has the effect of removing the mapping in sysmem. So
now the sysmem pager only does maps and doesn't try to do the other function's
job at all.

The atomic stuff was also removed; instead, exclusion on that hash table is
kinda-sorta achieved by setting the sysmem pager thread to a lower priority
than the launching thread. Properly we should exit the sysmem pager thread
before cleaning the hash table and its items, but whatever.