user/crt: parse argument buffers correctly
Also do the first call to brk() accordingly. This system is silly as noted in
the comment, and should be rewritten to lean on UAPI more heavily; but it does
work (for less than half a page's worth of either arguments or environment
variables), so further work can proceed on top.

For benchmark's sake, initrd_spawn_partner gets to VM::brk in ten faults. A
better solution would likely improve on this heavily, e.g. incurring faults on
argv and envp only when they're referenced within main().