user: add interface and basic tests for fork(2), wait(2)
These are, for the moment, dependent on L4.X2 IPC, but should be changed to
use plain POSIX stuff once e.g. the return code of the child process gets
brought back over wait(2).

There's also an iteration variable to test wait(2) in both passive and active
modes, i.e. regardless of which side gets to the rendezvous point first.