sys/vm: implement VM::fork
The way copy-on-write sharing works here is that physical pages have a primary
owner and, depending on the kind of virtual memory they're each attached to,
zero or more secondary owners. The primary is tracked in <struct pp> and
secondaries are indexed in share_table by their physical page number.
Consequently fault handling may copy when the faulter isn't the sole owner of
a page, and disable copy-on-writeness otherwise, while performing just one
hash table query to determine whether there are more than 1 or 2 secondaries,
depending on the ownership pointer.

In the future, replacement of shared physical memory will be implemented in
terms of the general phys-to-virt mapping that this use of share_table
provides. It may also be possible to rewrite this in terms of a
lock-free/wait-free hash table to yield lf/wf forking and page replacement.