sys/crt: add cnd_*() stubs, basic tests
This is part of the C11 <threads.h> interface. The header file is imported
from mung even when it should properly be copied instead. Oh well.

There are no tests for cnd_timedwait(), and the other two might not complete
if implementations are missing entirely. However, quarter-arsed is better than
nothing, so the former is what we get until there's need to do them properly.