sysmem: disable start_abend_helper()
It suffers from a massive circular dependency issue. To fix, root should
create the helper threads explicitly as part of the initialization protocol,
since the two are already conjoined at the hip and sysmem has no need to
create arbitrary numbers of threads for itself to begin with.

This is not a big loss since the whole "abend" protocol hasn't worked for
months now.