root: make the early sbrk() heap slightly more robust
It'll report trying to record too many fpages, which happens when root's image
ends at a sufficiently unaligned address to amplify the shrapnel factor, or
when the boot modules grow to greater sizes. The alternative option would've
coalesced fpages of equal size, but in the end it's better to increase the
startup static memory allocation than to add code.

Unrelatedly, add some asserts to sysmem's alloc_struct() because it's possible
and because there's no upcoming sysmem changeset it'd fit in with better.