user/crt: implement userspace part of sigaction()
Large chunks of the sigaction interface are missing, such as most of the flags
and quite a bit of the blocking behaviour. Most of which can't be tested until
signals signal signally signal response within signal handling code.

Also, sigaction()'s part where it sets ignore, default, or deliver disposition
is woefully undertested for the time being, but the asserts are solid enough.