root: add signal delivery in uapi.c
The part in sig_deliver() which creates a new thread, uses it to run a little
fragment of code in userspace, and then removes it on a condition in an IDL
routine, should be recast as a "threadlet" mechanism as soon as a second user
shows up.

Also, the catching part of uapi_sigset() should be something related to the
and_mask used to clear pending_flags, to allow smarter filtering without
causing signals to be lost because the user process doesn't maintain a
long-term copy of pending_set. But this also needs a test.