root: increase or remove various boot-time timeouts
There's no reasonable fallback path, so there's no point in taking errors for
most of these. The main rationale for a therefore-unfinished changeset is to
have all of this boot at least somewhat in QEMU on a KVM-less T43p.

Also, add some asserts about tno_free_maps prior to calls to
bitmap_{set,clear}_bit(). This was too small for a dedicated changeset, and
this one touches root/, so was a position of opportunity. Similar for error
report of PT_LOAD segment not aligned to page, in uapi.c .