user/crt: SA_NODEFER
The part where the signal itself is being masked unless SA_NODEFER is set.
However, if the signal is masked and then re-raised, it should be run
immediately after the previous handler completes, which this changeset makes
it do.

In particular, this makes block_set apply only to signals to be deferred under
sa_mask. Those that were will be reinvoked at the end of the __sig_invoke()
call that blocked them and therefore unblocks them at the end.