add conceal mode to Proc::sigset
Also called ``pending but obscured''. This is so that sigprocmask(2)'d signals
aren't delivered to __sig_bottom() and aren't removed from the pending set,
but are accessible to sigpending(2). The signal handling helper's terminating
condition was adjusted to trigger on the concealed set.

Unmasked signals are now handled synchronously before sigprocmask(2) returns
similarly to the way kill(getpid(), ...) works. On the downside this causes an
extra call to Proc, and likely isn't safe against concurrent signal helper
invocation.