fix unaligned linkage & dependent code
The background is that all of the Makefile stuff comes from mung, which is a
microkernel which therefore wouldn't be loaded in by demand paging. So its
Makefile, and those of its support stuff like mbiloader, can set things up at
non-aligned addresses. These are no good for mmap(2) unless adapted, which we
used to do with VM::mmap.

This changeset makes it so that the root task and boot modules are still
linked with -N, but anything that spawns into userspace under vm isn't.
vm_mmap() was changed to be pickier about its parameters, and uapi_spawn() to
complain about unaligned executables where capable.