lib: add portable stdio
That's to say, the common part that's sharable between the system and
userspace runtimes. Implemented in terms of fopencookie() because hey why not.

This changeset touches just about every vestigial C runtime in sneks, so the
interface step of a cleaner development cycle would wind up breaking various
things and thereby prevent tests from being run. As such we introduce the
fully-developed changes first and tests second, so that further development
can have nicer separation.