root: implement Proc::sigsuspend
Also a fix to sig_deliver(); it now queues signals up in the pending mask when
the helper thread was already created. As noted in a TODO, an existing
siginvoke threadlet could be interrupted by a second threadlet, but that's
hardly worth the trouble since all the threadlet does is interrupt the main
thread & put it in the pending-pumping loop.

[All of the signal handling stuff should move out of uapi.c and into a
proc_signal.c or some such. Similar with fork and spawn (and later execve),
and the UID stuff.]