sys: change sysmem's link address half a meg up
Ideally sysmem would be loaded into pages behind root's initial memory so that
root could grow during pre-sysmem init as it wanted, but for some reason
sigma0 won't respond to requests at 0x300000, so for now that idea is on the
backburner. This hack tides us over for the next 512 kilobyters of early root
memory.