user/crt: fix incorrect sigaction(2) idempotent case
When a SIG_IGN or SIG_DFL signal was put into the same disposition,
sigaction() would actually clear it using Proc::sigset. This was because the
routine previously ignored idempotent cases only in the first clause; with the
changes presented, ign/dfl bits are smashed only when the handler is set to a
non-ign/dfl disposition.