user/crt: don't lose signals when longjmp() exits handler
If a stack of N invocation frames are built on the stack for a group of N
signals, longjmp()ing out of a signal handler before the last would cause
loss of handler invocations for reliable signals. That's no good for e.g.
signal-driven I/O.

To repair this we change __sig_invoke() to a loop over an explicit processwide
signal queue, and intercept a longjmp() from within a signal handler such that
the first longjmp() takes effect after all signal handlers have run (or exited
through partially effective longjmp()). Similarly __sig_bottom() changes to
queue signals up if an existing handler chain would be interrupted.

This passes currently extant tests, but it's likely possible to construct
tests for POSIXly correct behaviour that'd break.