sys/vm: add longjmp() from sys/crt
Only under DEBUG_ME_HARDER though, which is where the other invariant stuff
comes in. This broke when longjmp() was divided into porcelain and plumbing a
few commits ago. It might be reasonable to add this as a weak symbol in lib/,
but since there's no immediately obvious place for it right now, this'll do.