sys/vm: remove anon+shared pages from page cache
Pages born into mmaps created MAP_ANONYMOUS | MAP_SHARED are found by owners
of that mmap through the page cache. When the final such mapping goes away,
those pages should be removed. Therefore the pagecache invariant check is
modified to test that at least one lazy_mmap exists for each pagecache page
served anonymously by vm.

To make that invariant check pass, we add an anon_mmap_table to track
lazy_mmaps that designate anonymous shared memory, and apply a brute-force
cut_anon_mmap() operator to the page cache as said mmaps are removed or
reshaped. In some callers <struct vp> removal was also moved up above the
first call to cut_anon_mmap() on the same region since it demands that
removable pages appear orphan.

More correctly we would release anon+shared pages as the page cache is subject
to ordinary page replacement. That's clearly most efficient as it'd get slower
as the number of mapped pages increases, and not with the space in potentia as
with the eagerly-reclaiming version.