sys/vm: move invariant checking to IPC quiet spots
Also change some comments around wrt the related support code in general.

This stuff is starting to take up a significant portion of vm.c, which
suggests that either it should be moved from vm.c to somewhere it isn't next
to the data structures which it concerns, or operations on said data
structures moved similarly. Either way vm.c's structs and variables are going
to move into a defs.h real soon now.