add systask selftesting
This adds, under BUILD_SELFTEST, Proc::get_systask_threads() which can be used
to enumerate all systasks for the purpose of executing them as part of regular
systests, and therefore having their outputs appear in the ordinary reports.

The rest is protocol and syntax to that end, and for executing tests in other
systasks and capturing their output and exit code in systest.