root: tap-ify threading selftests
This moves rootserv into a different thread, which is good because it should
have an ordinary stack, instead of the measly 4k it gets statically. Properly,
the main() thread could be removed by rootserv via IPC from after the waitmods
have completed, and its stack passed to sysmem, but that's a lot of fancy for
4k of RAM.

Also adds the runtime bits.