root: basic filesystem I/O
The fseek() thing is a bit half-arsed because there's no fstat() and no
Fs::seek call.

[This changeset alters the WIBNI spec of how the Sneks::Fs::read position
parameter works. See the new comment for what's up. It's good, but should
accept a signed 64-bit integer for position in order to break the 2 GiB
per-file limit (which we obviously want to do). Maybe muidl does this
already?]