add sys/msg.idl
This defines Sysmsg, effectively a thread-friendly "signals lite" for
systemspace, so that asynchronous events can be responded to in a manner more
suitable to native L4.X2 programs. The simplest application of this mechanism
is notification of process lifecycle, i.e. fork, exec, exit, across system
tasks that may hold per-client data.