sys/msg: remove filter when last label is removed
Per count only. This also resets all saturated buckets that may exist.

Also fix an inverted condition in filter_query(). This is just as embarrassing
as it sounds. However, the bug was found and repaired.

The two of these had an interesting effect where the second one caused a bug
(the filter was inverted) which the interface design bug concealed (a sister
test's leftover "reject all" filter, inverted, therefore behaved correctly in
test). This suggests that related tests should be designed together.