user/crt: add internal threads
These come with a bevy of limitations, such as not being allowed to call
malloc() (and by extension, printf()).

Apparently Proc::create_thread doesn't set a pager for the incoming thread.
What's up with that. Setting it at the call site thru ExchangeRegisters solves
the immediate problem; it could even be folded into the Start_SpIp to remove a
syscall.