user/crt: implement epoll edge-triggered mode
This is missing significant features such as level-triggered mode and
EPOLLONESHOT, and doesn't interact correctly with fork(), when epoll instances
are specified for polling in other epoll instances, or when multiple epoll
instances per process are listening for a given pair of fd-and-event. Trying
to catch SIGIO in userspace will conflict with epoll operation if one or more
epoll instance has been created in that process image (i.e. including its fork
parents).

In particular, forking puts all existing epoll instances in an undefined state
on the child process, recursive epoll does nothing and doesn't recognize
loops, and multiple epoll instances behave as though all were using
EPOLLEXCLUSIVE on every fd.

However it's good enough to do select(2) and poll(2) on top of, and that's
good enough for olde-schoole POSIX.