sys/chrdev: add <sneks/chrdev.h>, stubs
This is the interface, stubs, and Makefiles for libsneks-chrdev.a, which does
reusable client and file description management for services that behave like
a character device, i.e. a buffer with some combination of readers, writers,
and background tasks (such as hardware interactions). The library is designed
to also handle sysmsg interactions with regard to process forking, exec, and
termination. Effectively it conceals implementations of Sneks::IO,
Sneks::Poll, and Sneks::Pipe; in the future there will be interfaces for
opening character devices and named buffers (fifos, sockets) as well.

At time of writing this is directly applicable to implementation of POSIX
pipes and should also work for named pipes, bidirectional sockets, actual
character devices (serial ports etc.), and virtual character devices
(/dev/{[nf]ull,zero}). In particular it does this without reference to any
microkernel interaction or data type, so in principle its clients should be
compatible with equivalent knick-knacks in e.g. GNU/Hurd, given a suitable
reimplementation.

Its callback specification interface was designed as a homage to GLUT, but
should be modifiable to exploit link-time symbol resolution instead of
introducing a gnarly data-to-control dependency for each callback invocation.
Since there are no benchmarks for telling how much faster that would be, the
optimization is premature and therefore not specified here.

Currently missing are a fire-tested interface for ioctl(), since it remains
unspecified on the IDL level, and of course the implementation in a later
changeset. It's also unknown whether any implementor will benefit from the
dead_client interface, so that may be coming off in the near future.