sys/pipe: rework for <sneks/chrdev.h>
Removes a whole bunch of stuff that's being moved to and/or reimplemented in
sys/chrdev in subsequent changesets of this series. In the current commit
order, this links to mere stubs of the interface and therefore fails to
function.

Additionally, use chrdev's interface for confirming effects subject to erasure
upon IPC reply failure. This checks off a couple of TODOs in the old pipeserv,
however the IPC failure side of this function is currently untested.