fs.squashfs: implement Sneks::Directory
To this end, we add a dentry cache keyed by directory inode number and either
dentry name or position within its directory, and separate dentry reading out
of lookup(). The dentry cache has no replacement function; that can wait until
common filesystem operations are moved into a libsneks-fs.a or some such.

The dentry cache enables simple but effective getdents() and seekdir() albeit
with somewhat complex a routine for low-level dentry access. getdents() is a
bit clever in returning an increasing number of dentries per call to save on
very small queries, such as immediately following seekdir() or rewinddir().

In order to reduce duplication, this also reimplements lookup() in terms of
the dentry cache. As a result the filesystem should be a bit faster by
decompressing and/or interpreting fewer metadata blocks.

Other smaller changes:
  - use ino_t instead of uint64_t for inode numbers
  - remove old printf()s that were mostly commented out already
  - demote a few old FIXMEs to TODOs since they weren't critical
  - rework the last-part logic of squashfs_resolve() to work for directories