sys/io: define interface in <sneks/io.h>
This is a derivative of <sneks/chrdev.h> for the purpose of reusing a single
implementation of Sneks::IO and Sneks::Poll for handle management, lifecycle
handling, and whatever else will be shared between various filehandle
providers in the future. Like its predecessor it permits compatible versions
without reference to L4 types or semantics.