sys/io: first implementation of <sneks/io.h>
This is a big one.

Some parts are reused from elsewhere:
  - use of rangealloc to manage handles comes from squashfs.c
  - lifecycle message handling and main dispatch loop from chrdev.c
  - tracking of handle-to-file and client-to-handle relation from chrdev.c

Some are novel:
  - hash-and-flag implementation of the dup_to/touch pair
  - fast confirms to enable lazy close
  - invariant checks
  - shadow handles to support transparent forking

There is a limitation in that the maximum number of handles per implementor
task is USHRT_MAX - 1. This is so that <struct fd>'s rangealloc size is a
smaller multiple of 2. One way to manage the limitation would be to allocate
at most some number of handles for one client, half that number for another,
and so forth with the idea that a single task cannot DoS a given IO provider
by simply calling dup() in a loop.

Another way is based on the observation that since the limitation is actually
on the indexes into clients' and files' handle tracking structures, file
descriptors above a certain per-client count could be tracked not through
rangealloc but using a per-client hash table. This would only slow down a
handle-happy client, and its forked children, rather than allowing it to
impede overall system performance or causing it to error out.