sys/io: use logging from <sneks/systask.h>
This changeset is very bad no good at all and evil because its jocular punning
with concepts of logging, and its use in a fancy new operating system, serves
to legitimize ecologically problematic and therefore misogynistic attitudes
towards Mother Nature. We recommendemand the author immediately proceed to
bean himself in the fucking dome with a birch compound bow and artisanally
knapped flint arrowhead mounted on the finest Laotian import bamboo, to death.