fs.squashfs: lookups of current and parent directories
I.e. "." and "..".

Squashfs is a bit fucky in that it has two ways to designate an inode: one it
calls an "offset" which is actually a pair of block offset and position within
a block, and the other it calls an "inode_number" which is what conventional
filesystem designs would use. This complicates resolution of paths containing
".." because directory inodes give the parent in the "inode_number" format
from which there's no immediately obvious mapping to the block-and-position
form.

Luckily, squashfs directory entries contain both an entry's block-and-position
offset and its inode_number, and no file or directory is accessed without
reading its dentry first, so we're able to build an ino-to-blkoffset mapping
in read_dentry() and to synthesize valid dentries for "." and ".." in
directory positions 0 and 1 respectively.

Types were refined to use ino_t for inode numbers everywhere, and previously
confused usage of "ino" for the block-and-position form was renamed to
"blkoffset" in at least one place.