idl/api: support descriptor flags, FD_CLOEXEC in Sneks::IO
This splits IO::set_flags into IO::set_{handle,file}_flags so that both handle
("descriptor") flags and file ("description", "status") flags can be accessed.
Handle flags correspond to FD_* and file flags to O_* in POSIX.

The semantics of these calls were altered so that they no longer pop EINVAL
when previously unknown flags were set in or_mask or cleared in and_mask.
Instead serverside is expected to simply ignore those, and userspace can query
the actual result if it has an alternative for when a change didn't take.

This also adds the "flags" parameter to Sneks::IO/dup so that FD_CLOEXEC
(freshly defined for not already having been) can be set atomically on handle
creation. Implementors and bare-back callers of Sneks::IO/dup were updated to
add, and ignore, the new parameter.