user/test: create process:env
First two tests are stack_location and forbidden_zone. The first tests that
the runtime is correctly placing a process stack sized according to the
exec-time RLIMIT_STACK, i.e. that only a small stack may be located between 0
and the program image's low address.

The second tests for the presence and location of a forbidden zone at low
addresses, expecting to find the edge of one at exactly the 64k mark.