lib: move root/elf.h into include/sneks/ and update it
Update from glibc 2.32, modified to fit the tree and with AT_* defs moved into
<sys/auxv.h>.

The glibc <bits/auxv.h> is included from both <elf.h> and <sys/auxv.h>,
causing breakage because it's not ifndef-guarded. We solve this by removing
<bits/auxv.h> into <sys/auxv.h> and slapping _GNU_SOURCE around getauxval()
prototype. There's also a stub implementation of getauxval() for testing.