root: correct filesz/memsz confusion in load_program()
This is so that we can have file maps padded with zero on their last
file-backed page regardless of what's in the file after filesz ends, which is
how ELF binaries come out of the oven.

Simplify load_program() while we're here, and filter for PT_LOAD in
read_elf().