sys/vm: lose the brk bloom filter
What's this do, save a hash table lookup of a pre-existing page except when
it's full? When a bloom filter query is already a hash lookup, involving
further massage of the incoming hash value? When that hash table lookup is
just a bit of cache shuffling otherwise?

Lose that, k-rad bloom filter or not