lib: implement setbuf(3) family
A good number of various edge cases implied by the POSIX specs of these
functions are left untested, and therefore likely don't work. However, this
changeset does achieve printf() line atomicity wherever output should appear
through bootcon (which should identify as a teletype in a later isatty(3)
series), cleaning up some output of forked tests, if not the reporting
script's deficiencies.

fwrite() now has a crazy switch-case statement for dealing with line and block
buffering in an extra fancy cascading style that either reinvents what
primordial Unix did back in the 1970s, or is entirely novel.

File streams are now tracked and flushed on exit (and not flushed on _Exit,
though this is also not tested), and __stdio_file is no longer exported in
<stdio.h>.