sys/crt: add file table and lseek(2)
This corrects use of Sneks::Path/resolve by recording the server TID for each
file descriptor instead of always assuming they resolved to initrd_tid. It
also limits the number of free file descriptors to 13 per systask; this is
arbitrary and can be raised. Also, fileio.c's style is brought up to date.