sys/msg: refactor filtering semantics and style
Mask bits which a client has ever been subscribed to are added to the new
per-client filter_mask, and if a message's positive mask isn't entirely
present in all of a client's filter_mask, it'll be accepted outright without
consulting the bloom filter. There, that's more confusing than the actual
source code.

This also removes darrays in favour of hash tables, restyles everything else
to not suck as bad, adds for_each_bit(), and demands sausages everywhere.